Move icon/radiation/ to radiation/icon/
Convert relative import to absolute, update 35+ absolute import
references across radiation test files, icon_physics.py, and
parameters.py.
