+++ jcm/physics/icon/icon_physics.py
@@ -20,7 +20,7 @@ from jcm.physics.icon.constants import physical_constants
 from jcm.physics.icon.icon_coords import IconCoords
 
 # Import physics modules (will be implemented progressively)
-from jcm.physics.icon.radiation.radiation_scheme import radiation_scheme
+from jcm.physics.radiation.icon.radiation_scheme import radiation_scheme
 from jcm.physics.icon.icon_physics_data import RadiationData
 from jcm.physics.convection.icon import tiedtke_nordeng_convection
 from jcm.physics.clouds.icon import shallow_cloud_scheme, cloud_microphysics
@@ -835,7 +835,7 @@ def _apply_radiation_rrtmgp_inner(
     terrain: TerrainData
 ) -> tuple[PhysicsTendency, PhysicsData]:
     """Apply RRTMGP radiation heating rates (inner, always-compute version)."""
-    from jcm.physics.icon.radiation.radiation_scheme_rrtmgp import (
+    from jcm.physics.radiation.icon.radiation_scheme_rrtmgp import (
         radiation_scheme_rrtmgp,
     )
 
@@ -945,7 +945,7 @@ def _apply_radiation_emulated_inner(
     Uses bidirectional GRU networks to predict SW and LW fluxes for each
     atmospheric column, then derives heating rates from flux divergence.
     """
-    from jcm.physics.icon.radiation.radiation_scheme_emulated import (
+    from jcm.physics.radiation.icon.radiation_scheme_emulated import (
         radiation_scheme_emulated,
     )
