+++ jcm/physics/radiation/icon/__init__.py
